+++ scripts/scripts.js
@@ -16,6 +16,7 @@ import {
   decorateBlock,
   buildBlock,
 } from './lib-franklin.js';
+import enableStickyElements from './sticky-elements.js';
 
 /**
  * to add/remove a template, just add/remove it in the list below
@@ -24,8 +25,6 @@ const TEMPLATE_LIST = ['application-note', 'news', 'publication', 'blog'];
 
 const LCP_BLOCKS = ['hero']; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'molecular-devices'; // add your RUM generation information here
-
-let LAST_SCROLL_POSITION = 0;
 
 export function loadScript(url, callback, type, async) {
   const head = document.querySelector('head');
@@ -246,42 +245,6 @@ export async function fetchFragment(path) {
 }
 
 /**
- * Enable sticky components
- *
- */
-function enableStickyElements() {
-  const stickyElements = document.querySelectorAll('.sticky-element');
-  const offsets = [];
-
-  stickyElements.forEach((element, index) => {
-    offsets[index] = element.offsetTop;
-  });
-
-  window.addEventListener('scroll', () => {
-    const currentScrollPosition = window.pageYOffset;
-    let stackedHeight = 0;
-    stickyElements.forEach((element, index) => {
-      if (currentScrollPosition > offsets[index] - stackedHeight) {
-        element.classList.add('sticky');
-        element.style.top = `${stackedHeight}px`;
-        stackedHeight += element.offsetHeight;
-      } else {
-        element.classList.remove('sticky');
-        element.style.top = '';
-      }
-
-      if (currentScrollPosition < LAST_SCROLL_POSITION && currentScrollPosition <= offsets[index]) {
-        element.style.top = `${Math.max(offsets[index] - currentScrollPosition, stackedHeight - element.offsetHeight)}px`;
-      } else {
-        element.style.top = `${stackedHeight - element.offsetHeight}px`;
-      }
-    });
-
-    LAST_SCROLL_POSITION = currentScrollPosition;
-  });
-}
-
-/**
  * loads everything that doesn't need to be delayed.
  */
 async function loadLazy(doc) {
@@ -289,10 +252,9 @@ async function loadLazy(doc) {
 
   // eslint-disable-next-line no-unused-vars
   const headerBlock = loadHeader(doc.querySelector('header'));
-  // we run this twice to ensure that the sticky elements are enabled as soon as the header appears
-  enableStickyElements();
 
   await loadBlocks(main);
+
   enableStickyElements();
 
   const { hash } = window.location;
